backend/requirements.txt
g4f
Flask
docx2pdf
python-docx
selenium
pyperclip
curl_cffi
flask-cors